demo search user full text

+++ app/Http/Controllers/UserController.php
@@ -6,9 +6,8 @@ use App\Http\Requests\CreateUserRequest;
 use App\Http\Services\UserService;
 use App\Role;
 use App\User;
-use Carbon\Carbon;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -21,18 +20,27 @@ class UserController extends Controller
 
     public function index()
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $users = $this->userService->getAll();
         return view('admin.users.list', compact('users'));
     }
 
     public function create()
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403,'Khong co quyen');
+        }
         $roles = Role::all();
         return view('admin.users.add', compact('roles'));
     }
 
     public function store(CreateUserRequest $request)
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $this->userService->create($request);
         toastr()->success('create user success!');
         return redirect()->route('users.index');
@@ -40,6 +48,9 @@ class UserController extends Controller
 
     public function delete($id)
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $user = $this->userService->findById($id);
         $user->roles()->detach();
         $this->userService->delete($user);
@@ -49,12 +60,18 @@ class UserController extends Controller
 
     public function edit($id)
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $user = $this->userService->findById($id);
         return view('admin.users.edit', compact('user'));
     }
 
     public function update(Request $request, $id)
     {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $user = $this->userService->findById($id);
         $this->userService->update($user, $request);
         toastr()->success('update user success!');
@@ -71,7 +88,6 @@ class UserController extends Controller
     {
         $users = User::where('name','LIKE', '%' . $request->keyword . '%')
                         ->orWhere('email', 'LIKE', '%' . $request->keyword . '%')
-
             ->with('roles')->get();
 
         return response()->json($users);
